Add a debian/shlibs.local file.
Otherwise the build fails at dh_shlibdeps since there is no indication which
package(s) the libconnui* libs are in.

An alternative would be the run dh_makeshlibs in libconnui-dev but that
leads -- without further tweaks -- to the a wrong shlibs file (pointing to
libconnui-dev instead of instead of libconnui{,-cellular}).

So this approach is not very elegant but reasonable simple and it works.